refactor(VendingMachineController): controller 패키지로 이동
